requirements.txt
ultralytics>=8.0.0
streamlit>=1.20.0
opencv-python-headless>=4.5.5
Pillow>=9.0.0
numpy>=1.20.0
torch>=2.0.0
torchvision>=0.15.0
gdown>=5.1.0
